web01/requirements.txt
django==2.2.3
djangorestframework
